Adding Graphics Framework and Grimelins source
The graphics framework is something I made to have a geometry library (
MPR/Raycasting/Primitive Generation/Etc ) and math library in one place
along with a nice DX10 renderer example.  The Grimelins source is from a
year or two ago and is DX9, as well as a TON of other C++ code.